Move generators to separate module
