+++ tests/Peer.hs
@@ -1,3 +1,4 @@
+-- | Tests for the "client protocol".
 module Peer (tests) where
 
 import Protolude hiding (phase)
@@ -12,18 +13,18 @@ import qualified Crypto.Spake2.Group as Group
 import Crypto.Spake2.Groups (Ed25519(Ed25519))
 import qualified MagicWormhole.Internal.Peer as Peer
 
-import qualified Messages as MessagesTest
+import qualified Generator
 
 tests :: IO TestTree
 tests = pure $ testGroup "Peer"
   [ testProperty "SPAKE2 messages roundtrip" $ property $ do
-      appID <- forAll MessagesTest.appIDs
+      appID <- forAll Generator.appIDs
       let protocol = Peer.wormholeSpakeProtocol appID
       element <- forAll elements
       tripping element (Peer.encodeElement protocol) (Peer.decodeElement protocol)
   , testProperty "SecretBox encryption roundtrips" $ property $ do
-      phase <- forAll MessagesTest.phases
-      side <- forAll MessagesTest.sides
+      phase <- forAll Generator.phases
+      side <- forAll Generator.sides
       secret <- forAll $ Gen.bytes (Range.linear 1 10)
       let key = Peer.derivePhaseKey secret side phase
       plaintext <- forAll $ Gen.bytes (Range.linear 1 256)
